+++ CHANGELOG.md
@@ -22,13 +22,14 @@
 * fix(InputMenu/SelectMenu): sort filtered items by match relevance
 * fix(components): nullable and optional type support
 * chore(deps): update @nuxt/test-utils + vitest to v4
+* feat(dictionary): add menu icon
+* docs(CSS Variables): improve
 * feat(Theme): new component
 * feat(Footer): new component
 * feat(FooterColumns): new component
 * feat(Header): new component
 * feat(Main): new component
-* feat(dictionary): add menu icon
-* docs(CSS Variables): improve
+* feat(Drawer): new component
 
 ## [2.3.0](https://github.com/bitrix24/b24ui/compare/v2.2.1...v2.3.0) (2026-02-12)